Passivation

Documentation for Passivation.

Passivation.ispassive

Passivation.klap

Passivation.klap_inital_guess

Passivation.kyp

Passivation.kypare

Passivation.kypextremal

Passivation.kypmat

Passivation.kypmax

Passivation.kypmin

Passivation.lmi

Passivation.lmi_tp

Passivation.passivate

Passivation.popov

Passivation.popovplot

Passivation.sampopov

Passivation.ispassive — Method

ispassive(Σ; opt=:lmi, kwargs...)

Checks if the system Σ is passive. The function returns true if the system is passive and false otherwise. If opt is set to :lmi, the KYP lmi is tried to solve. If opt is set to :popov, the Popov function is sampled to check passivity.

Passivation.klap — Function

klap(Σ::StateSpace; L0=L0(Σ), M=M(Σ), P=gram(Σ, :c); recycl=:schur, restart=false, α=1e-8, ε=1e-4, verbose=true, kwargs...) -> Σp, res

Passivates a system Σ using the KLAP method. The optimization problem is solved using LBFGS.

Passivation.klap_inital_guess — Function

klap_inital_guess(Σ, ΔD=0.0; ε=0) -> L0, ΔD

Computes an initial guess for the KLAP optimization problem. The initial guess is computed by perturbing the feedthrough matrix to achieve a passive realization. Then the perturbed system is used to compute the initial guess. The perturbation ΔD can be specified, otherwise it is computed using ΔD(Σ).

Passivation.kyp — Method

X = kyp(Σ; optimizer=Hypatia.Optimizer, kwargs...)

Returns a solution to the KYP inequality by solving the corresponding linear matrix inequality.

Passivation.kypare — Method

W = kypare(Σ, X)

Returns the residual of the KYP ARE for some solution candidate X.

Passivation.kypextremal — Function

X = kypmax(Σ; kwargs...)

Returns the maximal or minimal solution to the KYP inequality by solving the Riccati equation for the stabilizing or anti-stabilizing solution.

Passivation.kypmat — Method

W = kypmat(Σ, X)

Returns the KYP matrix of the system Σ for the given matrix X.

Passivation.kypmax — Method

Xmin = kypmax(Σ; kwargs...)

Returns the maximal solution to the KYP inequality by solving the Riccati equation for the anti-stabilizing solution.

Passivation.kypmin — Method

Xmin = kypmin(Σ; kwargs...)

Returns the minimal solution to the KYP inequality by solving the Riccati equation for the stabilizing solution.

Passivation.lmi — Method

lmi(Σ::StateSpace; kwargs...) -> Σp, model

Solves the passivation problem for a given state-space system Σ using the positive real LMI constraints [1].

Passivation.lmi_tp — Method

lmi_tp(Σ::StateSpace; kwargs...) -> Σp, model

Solves the passivation problem for a given state-space system Σ using the positive real LMI constraints with trace parametrization [2, 3].

Passivation.passivate — Function

passivate(Σ::StateSpace, method=:klap, args...; kwargs...) -> Σp, res

Passivate the system Σ using the method method. The available methods are:

:klap: KLAP optimization

:lmi: LMI optimization [1]

:lmi_tp: LMI optimization with trace parametrization [2, 3]

The remaining arguments args and keyword arguments kwargs are passed to the corresponding passivation function.

Passivation.popov — Method

popov(Σ, s)
popov(Σ, ω)

Evaluates the popov function of the system Σ at the complex variable s or real frequencies ω.

Φ(s) = Σ(s) + Σ(-s)'

where Σ(s) is the frequency response (transfer function) of Σ at the complex variable s.

Passivation.popovplot — Method

popovplot(Σ; args...; kwargs...)
popovplot!(p, Σ; args...; kwargs...)

Plots the minimal eigenvalue of the Popov function of the system Σ at the frequencies ω.

Passivation.sampopov — Method

sampopov(Σ; ω=10 .^ range(-15, stop=5, length=5000)) -> Λmin, ω, Ωp, Ωnp, F, c

Samples the Popov function of the system Σ at the frequencies ω. Returns the minimum eigenvalues of the sampled Popov function, the frequencies ω, the passive and non-passive intervals Ωp and Ωnp, the eigendecompositions F and the change points c.

References

[1]

S. Grivet-Talocia and L. Silveira. Post-processing methods for passivity enforcement. In: Model Order Reduction: Volume 1 – System- and Data-Driven Methods and Algorithms, edited by P. Benner, S. Grivet-Talocia, A. Quarteroni, G. Rozza, W. Schilders and L. M. Silveira (De Gruyter, Berlin, Boston, 2021); pp. 139–180.

[2]

B. Dumitrescu. Parameterization of Positive-Real Transfer Functions with Fixed Poles. IEEE Trans. Circuits Syst. I. Regul. Pap. 49, 523–526 (2002).

[3]

C. Coelho, J. Phillips and L. Silveira. A Convex Programming Approach for Generating Guaranteed Passive Approximations to Tabulated Frequency-Data. IEEE Trans. Comput.-Aided Des. Integr. Circuits Syst. 23, 293–301 (2004).